refactor(tests): split llm mock responses
